[IMPROVE/DEL] Topic 및 Statistics 네트워크 통신 로직 개선 #3
