+++ PhotoExplorer/PhotoExplorer/Models/Entities/PhotoSearchRequest.swift
