+++ PhotoExplorer/PhotoExplorer/Models/Network/UnsplashRequest.swift
@@ -55,3 +55,4 @@ enum UnsplashRequest {
         }
     }
 }
+
